This patch allows fdb_getall_reverse to return undecoded line to filter.
This raw line can be filtered out earlier without the need of decoding
over the layout.

This is useful for large and long maintained projects. Projects
(Factory, Devel..)  have a long build history, decoding every
line in the jobhistory file only to be filtered out based on filter
conditions (package type, result code, buildtimes) is expensive. This
patch facilitates skipping those costs reach the result faster.

Tested on IBS/SUSE:FACTORY:HEAD/standard/i586/MozillaThunderbird
without patch:
100 results: 6 sec
1000 results: 53 sec

with patch:
100 results: 1 sec
1000 results: 13 sec

<!---
If you haven't done so already, please read the CONTRIBUTING.md file to learn
how we work and what we expect from all contributors.

https://github.com/openSUSE/open-build-service/blob/master/CONTRIBUTING.md

In order to make it as easy as possible for other developers to review your
pull request we ask you to:

- Explain what this PR is about in the description
- Explain the steps the reviewer has to follow to verify your change
- If the reviewer needs sample data to verify your change, please explain how to
  create that data
- If you include visual changes in this PR, please add screenshots or GIFs
- If you address performance in this PR, add benchmark data or explain how the
  reviewer can benchmark this

This is a good PR description example:

Hey Friends,

this introduces labels for the different build result states on the project
monitor page. This makes it easier to get a visual overview of what is going on
in your project.

To verify this feature

- Enable the interconnect to build.opensuse.org
- Create the project home:Admin
- Add 'openSUSE Tumbleweed' as a repository to the project
- Branch a couple of packages into the project:
  ```
  for i in `osc -A http://0.0.0.0:3000 ls openSUSE.org:home:hennevogel`; do osc -A http://0.0.0.0:3000 copypac openSUSE.org:home:hennevogel $i home:Admin; done
  ```
- Visit the monitor page and see the new labels for the different states.

Here is a screenshot of how it looks:

** Before **
![Screenshot of the project monitor](https://example.com/screenshot1.png)

** After **
![Screenshot of the project monitor](https://example.com/screenshot2.png)

-->