@mlschroe 2020-05-22
So we need FileDB::encode_field:
```
sub encode_field {
  my ($str) = @_;
  $str =~ s/([\000-\037%|=\177-\237])/sprintf("%%%02X", ord($1))/ge;
  return $str;
}
```

With that we do:
```
  my %packid = map {BSFileDB($_) => 1} @{$cgi->{'package'}};
```
and with that we do the matching in the filter.