@sjamgade 2020-05-28
Merging as discussed in standup